requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
fastapi
uvicorn
pydantic
joblib
matplotlib
seaborn
optuna
xgboost
imblearn